==> notebooks/00_eda.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

dataset = pd.read_csv('../data/raw/online_retail_II.csv')

dataset.head()

dataset.columns

dataset.info()

dataset[dataset.Description.isna()]

new_dataset = dataset.copy()

new_dataset.columns

new_dataset.drop(columns=['Description' , 'Customer ID' , 'StockCode'] , inplace= True)
new_dataset.columns

new_dataset.isna().sum()

new_dataset['TotalAmount'] = new_dataset['Quantity'] * new_dataset['Price']

new_dataset

quantile = new_dataset['TotalAmount'].quantile(0.95)
quantile

new_dataset

dates = new_dataset.InvoiceDate
dates

def get_date(df, column_name):
    df[column_name] = pd.to_datetime(df[column_name])
    
    df['Year']  = df[column_name].dt.year
    df['Month'] = df[column_name].dt.month
    df['Day']   = df[column_name].dt.day
    df['Hour']  = df[column_name].dt.hour
    df['min'] = df[column_name].dt.minute
    df['sec'] = df[column_name].dt.second
    
    return df

new_dataset = get_date(new_dataset, 'InvoiceDate')

print(new_dataset[['InvoiceDate', 'Year', 'Month', 'Day', 'Hour' , 'min' , 'sec']].head())

new_dataset.columns

new_dataset.drop(columns=['Invoice'] , inplace= True)

new_dataset.columns

new_dataset.Country.unique()

from sklearn.preprocessing import LabelEncoder

le = LabelEncoder()

new_dataset['Country'] = le.fit_transform(new_dataset['Country'])

new_dataset

new_dataset.info()

new_dataset['value'] = np.where(new_dataset['TotalAmount'] > quantile, 1, 0)

new_dataset.drop(columns=['Invoice' , 'TotalAmount'] , inplace= True)

new_dataset

new_dataset.to_csv('../data/processed/data.csv', index=False)
